# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me/orders/{id}": {
      "get": {
        "operationId": "UsersController_getOrderDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/rooms/{id}/order/{orderId}": {
      "post": {
        "operationId": "RoomsController_linkOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rooms"
        ]
      }
    },
...
    "/orders": {
      "get": {
        "operationId": "OrdersController_getMyOrders",
        "summary": "Get current user's orders",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getOrderDetails",
        "summary": "Get order details by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/checkout": {
      "post": {
        "operationId": "OrdersController_checkout",
        "summary": "Complete checkout (Create Order)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/confirm": {
      "post": {
        "operationId": "OrdersController_confirm",
        "summary": "Confirm Razorpay payment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/{id}/gift": {
      "post": {
        "operationId": "OrdersController_addGift",
        "summary": "Add a free gift to an eligible order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/create": {
      "post": {
        "operationId": "OrdersController_createOrder",
        "summary": "Create direct order (e.g. COD)",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/admin/all": {
      "get": {
        "operationId": "OrdersController_getAllAdminOrders",
        "summary": "Admin: Get all orders",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/admin/{id}": {
      "get": {
        "operationId": "OrdersController_getAdminOrderDetails",
        "summary": "Admin: Get order details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/admin/pos/orders": {
      "post": {
        "operationId": "OrdersController_createPosOrder",
        "summary": "Admin: Create POS order",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/{id}/cancel": {
      "post": {
        "operationId": "OrdersController_cancelOrder",
        "summary": "Cancel an order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "example": "Changed mind"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "operationId": "OrdersController_updateStatus",
        "summary": "Update order status (Admin/Vendor)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "SHIPPED",
                    "enum": [
                      "PACKED",
                      "SHIPPED",
                      "DELIVERED",
                      "PAID"
                    ]
                  },
                  "notes": {
                    "type": "string",
                    "example": "Shipped via BlueDart"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/{id}/tracking": {
      "get": {
        "operationId": "OrdersController_getTracking",
        "summary": "Get order tracking info",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/admin/orders/{id}": {
      "get": {
        "operationId": "OrdersAdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/orders/{id}/status": {
      "patch": {
        "operationId": "OrdersAdminController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/checkout": {
      "post": {
        "operationId": "CheckoutController_checkout",
        "summary": "Process checkout (COD or ONLINE)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully"
          },
          "400": {
            "description": "Bad Request (Empty Cart, Stock Issue)"
          }
        },
        "tags": [
          "Checkout"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "RISBOW API",
    "description": "\n# RISBOW Ecommerce Super App API\n\nWelcome to the RISBOW API documentation!\n\n## Key Features:\n- 🔐 **Authentication** - OTP & Email/Password login\n- 🛒 **Orders** - Complete order lifecycle\n- 📦 **Products** - Rich catalog\n- 💳 **Payments** - Razorpay integration\n- 🎁 **Promotions** - Gifts & Coupons\n- 👥 **Rooms** - Live shopping\n\n## Usage:\nRefer to 'API_INTEGRATION_GUIDE.md' for flow details.\n        ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://api.risbow.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your JWT token from /auth/login",
        "in": "header"
      }
    },
    "schemas": {
...
      "CheckoutDto": {
        "type": "object",
        "properties": {
          "paymentMode": {
            "type": "string",
            "enum": [
              "COD",
              "ONLINE"
            ],
            "example": "ONLINE",
            "description": "Payment mode: ONLINE (Razorpay) or COD"
          },
          "shippingAddressId": {
            "type": "string",
            "example": "addr_123456",
            "description": "ID of the shipping address"
          },
          "notes": {
            "type": "string",
            "example": "Leave at front door",
            "description": "Optional delivery notes"
          },
          "giftId": {
            "type": "string",
            "example": "gift_123456",
            "description": "Optional gift SKU ID to include with order"
          },
          "couponCode": {
            "type": "string",
            "example": "SAVE50",
            "description": "Optional coupon code to apply discount"
          }
        },
        "required": [
          "paymentMode",
          "shippingAddressId"
        ]
      },
      "ConfirmOrderDto": {
        "type": "object",
        "properties": {
          "razorpayOrderId": {
            "type": "string",
            "example": "order_rcl..."
          },
          "razorpayPaymentId": {
            "type": "string",
            "example": "pay_..."
          },
          "razorpaySignature": {
            "type": "string",
            "example": "signature_hash"
          }
        },
        "required": [
          "razorpayOrderId",
          "razorpayPaymentId",
          "razorpaySignature"
        ]
      },
...
    }
  }
}